Issue #14: rename validate_clerk_input.py to validation.py
